Implementing style in Login and Signup
