+++ src/views/index.js
@@ -1,0 +1,11 @@
+import App from './App';
+import BoardAdmin from './BoardAdmin';
+import Login from './Login';
+import SignUp from './SignUp';
+
+export {
+  App,
+  BoardAdmin,
+  Login,
+  SignUp,
+};
